+++ src/main/java/nl/esciencecenter/praline/integeralign/AffineGapCost.java
@@ -1,17 +1,17 @@
 package nl.esciencecenter.praline.integeralign;
 
 public class AffineGapCost implements IGapCost{
-    final int start;
-    final int extend;
+    final float start;
+    final float extend;
 
-    public AffineGapCost(int start, int extend) {
+    public AffineGapCost(float start, float extend) {
         this.start = start;
         this.extend = extend;
     }
 
 
     @Override
-    public int getGapCost(int gapSize) {
+    public float getGapCost(int gapSize) {
         if(gapSize == 0) return 0;
         return start + (gapSize - 1) * extend;
     }
