+++ src/main/java/nl/esciencecenter/praline/integeralign/IGapCost.java
@@ -3,5 +3,5 @@ package nl.esciencecenter.praline.integeralign;
 
 public interface IGapCost {
 
-    int getGapCost(int gapSize);
+    float getGapCost(int gapSize);
 }
