+++ src/main/java/nl/esciencecenter/praline/newalign/IPositionCost.java
@@ -1,5 +1,5 @@
 package nl.esciencecenter.praline.newalign;
 
 public interface IPositionCost {
-    double cost(int posA, int posB);
+    float cost(int posA, int posB);
 }
