+++ src/main/java/nl/esciencecenter/praline/newalign/LinearGapCost.java
@@ -11,7 +11,7 @@ public class LinearGapCost implements IGapCost {
     }
 
     @Override
-    public double getGapCost(int gapSize) {
+    public float getGapCost(int gapSize) {
         return gapSize * cost;
     }
 }
